api: return opaque engine handle in `onEngineRunning`
To support interop between the Swift/Kotlin and C++ APIs.

+++ library/swift/EngineBuilder.swift
@@ -39,7 +39,7 @@ open class EngineBuilder: NSObject {
   private var appVersion: String = "unspecified"
   private var appId: String = "unspecified"
   private var virtualClusters: String = "[]"
-  private var onEngineRunning: (() -> Void)?
+  private var onEngineRunning: ((Int) -> Void)?
   private var logger: ((String) -> Void)?
   private var eventTracker: (([String: String]) -> Void)?
   private(set) var monitoringMode: NetworkMonitoringMode = .pathMonitor
@@ -394,11 +394,12 @@ open class EngineBuilder: NSObject {
 
   /// Set a closure to be called when the engine finishes its async startup and begins running.
   ///
-  /// - parameter closure: The closure to be called.
+  /// - parameter closure: The closure to be called. Passes an engine handle that can be used
+  ///                      to interoperate with Envoy's C++ APIs.
   ///
   /// - returns: This builder.
   @discardableResult
-  public func setOnEngineRunning(closure: @escaping () -> Void) -> Self {
+  public func setOnEngineRunning(closure: @escaping (Int) -> Void) -> Self {
     self.onEngineRunning = closure
     return self
   }
